command-center-dashboard/backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
starlette==0.27.0

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
redis==5.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.2

# WebSocket
python-socketio==5.10.0
python-socketio[client]==5.10.0

# Data processing
pandas==2.1.4
numpy==1.25.2
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP client
httpx==0.25.2
aiohttp==3.9.1

# Background tasks
celery==5.3.4
redis==5.0.1

# Monitoring
prometheus-client==0.19.0
structlog==23.2.0

# Environment
python-dotenv==1.0.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Date/time
python-dateutil==2.8.2

# Cryptography
cryptography==41.0.7

# CORS
python-cors==1.0.1